Basic Tail Turn Implementation
By saving the movement string into a new list "_snakeMoveData" I create a pseudo array of vectors. I will save the movement input, which is also the key for the dictionary "_headDirection". I can then feed these movement 'arrays' down to the tail and then rotate them based of the dictionaries paired values. As long as the frame/sprite is oriented the same was as the source face sprite, then there doesn't need to be any alteration.

Some other minor cleanup.